Adding files to generate rof->ocn mapping files
JRA_runoff_carib.ipynb is an example of adding a mask
to the JRA-55 runoff dataset to represent the domain of the
Caribbean regional configuration.

JRA_runoff_domain.ncl is an example for creating a domain file
for the runoff dataset (JRA-55 TL319 in this example) and include
the mask created above to represent the domain of the regional
configuration. It is necessary to add this mask to avoid rivers
being placed in the wrong location. By masking out the region outside
of the regional domain, we assure that only the rivers within the
regional domain are considered during the mapping process.

map_JRA_to_Carib_025.nml is an example of namelist file to be used
with the runoff_map tool.